Nodes/requirements.txt
Pyro4>=4.82
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
psutil>=5.8.0